chore: joined.tsx와 certifed-challenges.tsx에서 동일한 컴포넌트를 return하도록 수정
